+++ src/headerdef.rs
@@ -63,6 +63,11 @@ pub enum HeaderDef {
     Sender,
     EphemeralTimer,
     Received,
+
+    /// A header that includes the results of the DKIM, SPF and DMARC checks.
+    /// See <https://datatracker.ietf.org/doc/html/rfc8601>
+    AuthenticationResults,
+
     _TestHeader,
 }